Rename ballot
